% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{gh_decode}
\alias{gh_decode}
\title{GeoHex V3 decode function}
\usage{
gh_decode(code)
}
\arguments{
\item{code}{Valid GeoHex V3 string}
}
\description{
This function takes a GH V3 code, then return lat,lon and level
}
\examples{
gh_decode('bb3371844188')
}
\keyword{geohex}

