% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{gh_encode}
\alias{gh_encode}
\title{GeoHex V3 encode function}
\usage{
gh_encode(lat, lon, level)
}
\arguments{
\item{lat}{Latitude
lon Longitude
level GeoHex level}
}
\description{
This function takes a lat,lon, level and returns GeoHex V3 code
}
\examples{
gh_encode(85.05112507763846, 89.37952995300293, 15)
}
\keyword{geohex}

